requirements.txt
streamlit==1.32.0
Pillow==10.2.0
pandas==2.2.3
openpyxl==3.1.2 